requirements.txt
fastapi
fastapi-sqlalchemy
pydantic
alembic
psycopg2
uvicorn
python-dotenv
pandas
Jinja2==3.0.3
python-multipart
numpy
Pillow
matplotlib
tensorflow